requirements.txt
Flask==2.2.3
imageai==3.0.3
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
plotly==5.13.1
pytz==2022.7.1
pytz==2022.7
tensorflow==2.11.0
tensorflow_intel==2.11.0
Werkzeug==2.2.3
